Added Functional Settings Tab
Added an Input Field in the Settings Tab along with a Save and a Back button.
When the Save Button is clicked it overwrites exerciceOne with the text written in the input field.

Had a problem with LandingView, it used to create a new instance of DataManager every time, so it reset every variables in there.
I found a workaround by simply creating a StateObject and referring it instead of the file DataManager() directly.